Avoid an "asymptotic space leak", aka excessive buffering
Lesson learned: It is not a good idea to simply capture a
std::shared_ptr and hope for the std::async implementation to actually
delete the lambda in a timely manner. It seems that at least my
implementation of std::async happily keeps a couple of function objects
around. Solve the apparent "space leak" by using a unique_ptr and
explicitly freeing the object.