silence a warning, add a FIXME
n_passes is a constexpr and does not requires any lambda capture - thus
remove it from the capture clause. Unfortunately, this makes gcc-8
unhappy because the compiler mistakenly diagonoses "n_passes" not to be
a constexpr in this context. Let's just remove the "constexpr" that way
gcc-8 compiles and clang-10 doesn't emit a warning that the capture is
superfluous.