Workaround: periodic b.c.: check for stride level consistency
The elimination process of periodic boundary conditions is not terribly
stable. It might happen that the order of two opposing degrees of
freedom (that are periodically constrained) flips after renumbering and
that we end up with the wrong degree of freedom (in the wrong stride) to
be eliminated. Add a bit of logic that detects this case and tries to
reorganize strides to make them consistent again.