Bugfix: compute a meaningful relaxation window instead of running into failsafe
Back in the beginning the `accumulate()` function was only called on
off-diagonal elements. But in order to fix an issue with the bounds on
the specific entropy (where we need the diagonal element itself).

But we now have to take the absolute value in the denominator in order
to not divide by zero.