EulerAEOS: remove the pressure function
Providing a `View::pressure(const state_type &U)` function is dangerous.
It creates the wrong perception that it is OK to call this function
freely when needed, while it is not... We obtain the pressure from the
oracle in a precomputation step and this is the only place from where
the pressure should be queried.

If absolutely necessary then the underlying equation of state oracles
can be called via
```
HyperbolicSystem::eos_pressure(rho, e)
HyperbolicSystem::eos_specific_internal_energy(rho, p)
HyperbolicSystem::eos_sound_speed(rho, e)
```