Local refinement: make index matching in SparsityPatternSIMD more robust
The communication pattern logic in SparsityPatternSIMD::reinit() was
written with the assumption that `partitioner->ghost_targets()` and
`partitioner->import_targets()` have a symmetry property. Namely, that
we import from and export to the very same set of MPI ranks. This
property might no longer hold true for locally refined meshes with
hanging nodes. Thus, we need to be more careful in matching indices when
constructing `indices_to_be_sent`.