SparseMatrixSIMD:: return a scalar or tensor with get_entry()
This change allows to use get_entry() also for tensor-valued matrices.
We now simply return a tensor in this case.

This allows for significantly more flexibility when implementing
limiters (and related) operations by allowing to switch from returning a
single scalar value to returning a (per component) tensor without
modifying the HyperbolicModule.