HyperbolicModule: count timings of synchronization barriers correctly
The logical or that we perform in order to synchronize the
"restart_needed" boolean is indeed a synchronization barrier. We thus
have to synchronization barriers in the explicit Euler step: one for
figuring out tau max and one for figuring out whether something signaled
a problem with the update step (and depending on policy we would need to
restart).