SparsityPatternSIMD: Bugfix: make MPI rank matching more robust
With locally refined meshes and hanging node constraints we might end up
in a situation where an MPI rank in the import_targets set does not show
up in the ghost_targets set. This is not a problem as we "minimize" the
sparsity pattern in (ghosted) locally owned rows to only contain entries
that we have in the locally relevant ranges of the tansposed locally
owned rows.

BUT, for the next MPI rank we have to try to match up from index 0
again.