Rename option to 'EXPENSIVE_BOUNDS_CHECK'
Let's fix this misnomer - we are always checking for IDP properties of
the low order update and signal a restart condition if the property is
violated. This configuration option merely enables additional more
expensive bounds checks that are normally not required. Thus rename the
configuration option to EXPENSIVE_BOUNDS_CHECK to make this clear.